CEDS-1238 Organise the code
